DeleteFilesService
