+++ src/integrationTest/java/uk/gov/hmcts/reform/bulkscanprocessor/tasks/DeleteCompleteFilesTaskTest.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import uk.gov.hmcts.reform.bulkscanprocessor.config.IntegrationTest;
 import uk.gov.hmcts.reform.bulkscanprocessor.entity.Envelope;
 import uk.gov.hmcts.reform.bulkscanprocessor.entity.EnvelopeRepository;
+import uk.gov.hmcts.reform.bulkscanprocessor.services.DeleteFilesService;
 import uk.gov.hmcts.reform.bulkscanprocessor.services.storage.LeaseAcquirer;
 import uk.gov.hmcts.reform.bulkscanprocessor.tasks.processor.BlobManager;
 import uk.gov.hmcts.reform.bulkscanprocessor.tasks.processor.EnvelopeProcessor;
@@ -42,6 +43,8 @@ public class DeleteCompleteFilesTaskTest {
     @Autowired
     private EnvelopeProcessor envelopeProcessor;
 
+    private DeleteFilesService deleteFilesService;
+
     @Mock
     private LeaseAcquirer leaseAcquirer;
 
@@ -49,10 +52,13 @@ public class DeleteCompleteFilesTaskTest {
 
     @BeforeEach
     public void setUp() {
-        this.task = new DeleteCompleteFilesTask(
-            blobManager,
+        deleteFilesService = new DeleteFilesService(
             envelopeRepository,
             leaseAcquirer
+        );
+        task = new DeleteCompleteFilesTask(
+            blobManager,
+            deleteFilesService
         );
     }
